cleanup: isolate query collector layout lowering
